requirements.txt
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
pymongo==3.12.0
python-dotenv==0.19.0
django-cors-headers==3.5.0
requests
gunicorn
